refactor(ensnode-sdk): move managed-names into sdk; delete redundant subregistry helpers
- New `packages/ensnode-sdk/src/shared/managed-names.ts` holds the
  namespace-parameterized `getManagedName(namespace, contract)` and
  `isNameWrapper(namespace, contract)` plus the `CONTRACTS_BY_MANAGED_NAME`
  structure (memoized per namespace). `apps/ensindexer/src/lib/managed-names.ts`
  becomes a thin wrapper that partially applies `config.namespace`.
- `getRootRegistryIds` now uses `getManagedName` directly (dropping the earlier
  mocked cross-package dependency).
- Delete `packages/ensnode-sdk/src/registrars/{basenames,lineanames,ethnames}-subregistry.ts`
  and drop their barrel exports. The sole consumer, `get-indexed-subregistries.ts`,
  is rewritten to iterate (plugin × datasource) entries and derive managed-name
  nodes via `getManagedName` — single source of truth.
