+++ packages/ensnode-sdk/src/registrars/index.ts
@@ -1,7 +1,4 @@
-export * from "./basenames-subregistry";
 export * from "./encoded-referrer";
-export * from "./ethnames-subregistry";
-export * from "./lineanames-subregistry";
 export * from "./registrar-action";
 export * from "./registration-expiration";
 export * from "./registration-lifecycle";
